requirements.txt
awswrangler==3.15.1
boto3==1.42.81
botocore==1.42.81
jmespath==1.1.0
numpy==2.4.4
packaging==26.0
pandas==2.3.3
pyarrow==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
s3transfer==0.16.0
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
